Diverging from java version: Use unsigned numbers to represent unsigned values

+++ src/Spice86/Emulator/Callback/ICallback.cs
@@ -2,5 +2,5 @@
 
 public interface ICallback : IRunnable {
 
-    public int GetIndex();
+    public byte GetIndex();
 }
